refactor: Create service for User creation
