+++ src/app/controllers/UserController.ts
@@ -2,12 +2,11 @@ import { Request, Response } from "express";
 import { StatusCodes } from "http-status-codes";
 import { ZodError, z } from "zod";
 
-import User from "../models/User";
-import { AppDataSource } from "../../database";
 import { GetAllUsersService } from "../services/GetAllUsersService";
+import { CreateUserService } from "../services/CreateUserService";
 
 const createUserSchema = z.object({
-  name: z.string(),
+  name: z.string().nonempty(),
   email: z.string().email(),
   password: z.string(),
 });
@@ -15,20 +14,10 @@ const createUserSchema = z.object({
 export class UserController {
   async create(req: Request, res: Response) {
     try {
-      const repository = AppDataSource.getRepository(User);
       const { name, email, password } = createUserSchema.parse(req.body);
 
-      const userExists = await repository.findOne({ where: { email } });
-
-      if (userExists) {
-        return res
-          .status(StatusCodes.BAD_REQUEST)
-          .json({ message: "Email already exists" });
-      }
-
-      const user = repository.create({ name, email, password });
-
-      await repository.save(user);
+      const service = new CreateUserService();
+      const user = await service.execute({ name, email, password });
 
       return res.status(StatusCodes.CREATED).json({
         id: user.id,
@@ -40,11 +29,16 @@ export class UserController {
           .status(StatusCodes.BAD_REQUEST)
           .json({ message: "Validation error", errors: e.issues });
       }
+      if (e instanceof Error) {
+        return res.status(StatusCodes.BAD_REQUEST).json({ message: e.message });
+      }
+
       return res
         .status(StatusCodes.INTERNAL_SERVER_ERROR)
         .json({ message: "Internal server error" });
     }
   }
+
   async list(req: Request, res: Response) {
     try {
       const search = req.query.search as string;
